Mateus

+++ Item.hpp
@@ -2,27 +2,27 @@
 #include <map>
 #include <string>
 
-class Item{
+class Item
+{
 public:
-
-    Item(int id, std::string name, float price, bool status);
+    Item(int id, std::string name, double price, bool status);
     std::string getName();
 
     /// @brief retorna o preço de uma unidade daquele item
     /// @param name nome do item buscado
-    float getUnitaryPrice();
+    double getUnitaryPrice();
 
     int getId();
 
     bool getStatus();
 
-    void changePrice(float novo);
+    void changePrice(double novo);
 
-    void changeStatus(); 
-      
+    void changeStatus();
+
 private:
     int _id;
     std::string _name;
-    float _price;
+    double _price;
     bool _status;
 };
